update agentic rag notebook
